WIP - Add Dag index Sentinel 2
Thie commit includes code designed to bulk import and import Sentinel 2
scenes from S3.

Given a year, month, and day this task will search for new tiles
uploaded for that given day. For each tile path found, an import job is
run that does the following:
 - creates a thumbnail object using the provided `preview.png` from
   sentinel 2
 - extracts the footprint and reprojects to 3857 based on information in
   `tileInfo.json`
 - create image objects based with appropriate metadata/band information
 - creates scene objects based on `tileInfo.json`

There are still a few tasks left to do with this:
 - [] verify that interaction with the API works
 - [] add the abiilty to poast thumbnails, images, and footprint during
 scene creation to streamline interaction
 - [] add documentation
 - [] verify backfilling works correctly and make job dynamic